pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emotivphysicimu"
version = "0.1.0"
description = "Feature extraction and light regression tools for Emotiv EEG/IMU artefact cleaning."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mapie",
    "matplotlib",
    "mne",
    "numpy",
    "pydantic",
    "python-dotenv",
    "pyyaml",
    "scikit-learn",
    "scipy",
]

[project.optional-dependencies]
catboost = ["catboost"]
dev = ["pytest"]

[project.scripts]
emotivphysicimu = "emotivphysicimu.cli:main"
